requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==25.9.1
google-auth==2.30.0
idna==3.7
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
